.claude/skills/ecg-analyzer/requirements.txt
# ECG分析器依赖包

# PDF处理
PyMuPDF==1.23.8
pdfplumber==0.9.0

# 科学计算
numpy==1.24.3
scipy==1.11.1

# 数据处理
pandas==2.0.3

# 可视化（可选）
matplotlib==3.7.1
plotly==5.15.0

# 图像处理（可选）
opencv-python==4.8.0.74
Pillow==10.0.0

# 其他
python-dateutil==2.8.2